requirements.txt
nipype==0.14.0
scipy==1.0.0
scikit_image==0.13.1
pydicom==0.9.9
itk_segmentation==4.12.0.post1
tensorflow_gpu==1.4.0
opencv_python==3.3.0.10
tensorlayer==1.7.3
matplotlib==2.1.0
numpy==1.13.3
SimpleITK==1.0.1
dicom==0.9.9.post1
itk==4.13.0
progressbar2==3.34.3
skimage==0.0
tensorflow==1.4.1
